move dataset classes to datasets.__init__
