+++ src/ranking_utils/datasets/antique.py
@@ -4,7 +4,7 @@ from collections import defaultdict
 from pathlib import Path
 from typing import Dict, Iterable, Set, Tuple
 
-from ranking_utils.dataset import ParsableDataset
+from ranking_utils.datasets import ParsableDataset
 
 
 class ANTIQUE(ParsableDataset):
