+++ src/ranking_utils/datasets/insuranceqa.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 from pathlib import Path
 from typing import Dict, Iterable, Set, Tuple
 
-from ranking_utils.dataset import ParsableDataset
+from ranking_utils.datasets import ParsableDataset
 
 
 class InsuranceQA(ParsableDataset):
