+++ src/ranking_utils/datasets/trec.py
@@ -3,7 +3,7 @@ from collections import defaultdict
 from pathlib import Path
 from typing import Dict, Iterable, Set, Tuple
 
-from ranking_utils.dataset import ParsableDataset
+from ranking_utils.datasets import ParsableDataset
 
 
 def read_qrels_trec(fname: Path) -> Dict[str, Dict[str, int]]:
